+++ src/main/java/com/switchfully/order/services/dto/ItemDTO.java
@@ -6,10 +6,10 @@ public class ItemDTO {
     private final String id;
     private final String name;
     private final String description;
-    private final BigDecimal price;
+    private final int price;
     private final int amount;
 
-    private ItemDTO(String id, String name, String description, BigDecimal price, int amount) {
+    private ItemDTO(String id, String name, String description, int price, int amount) {
         this.id = id;
         this.name = name;
         this.description = description;
@@ -29,7 +29,7 @@ public class ItemDTO {
         return description;
     }
 
-    public BigDecimal getPrice() {
+    public int getPrice() {
         return price;
     }
 
@@ -41,7 +41,7 @@ public class ItemDTO {
         private String id;
         private String name;
         private String description;
-        private BigDecimal price;
+        private int price;
         private int amount;
 
         public ItemDTOBuilder withId(String id) {
@@ -59,7 +59,7 @@ public class ItemDTO {
             return this;
         }
 
-        public ItemDTOBuilder withPrice(BigDecimal price) {
+        public ItemDTOBuilder withPrice(int price) {
             this.price = price;
             return this;
         }
